requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
libsql==0.1.11
libsql-experimental==0.0.55
mariadb==1.1.12
MarkupSafe==3.0.2
packaging==24.2
python-dotenv==1.2.1
SQLAlchemy==2.0.38
sqlalchemy-libsql==0.2.0
typing_extensions==4.12.2
Werkzeug==3.1.3
